requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
supabase
cryptography
celery[redis]
redis
openai
requests
